requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bayesian-optimization=1.1.0
hyperopt=0.2.5
ipython=7.22.0
pytorch=1.8.1
joblib=1.0.1
matplotlib=3.3.4
networkx=2.3
numpy=1.20.2
pandas=1.2.4
python=3.9.4
rdkit=2021.03.1
scikit-learn=0.24.1
scipy=1.6.2
statsmodels=0.12.2